Implement true support for FineTuneListEvents stream=true parameter
Break the API to two functions:

Task<FineTuneListEventsResponse> ListFineTuneEvents(string fineTuneId, CancellationToken cancellationToken = default)
non-streaming version that stops with all the events already generated.

IAsyncEnumerable<EventResponse> ListFineTuneEventsStream(string fineTuneId, CancellationToken cancellationToken = default) -
streaming version that blocks until a the next event is generated.  It completes when there is no more events generated.

The streaming version is complex due to the fact that the request can fail with a IOException "The response ended prematurely" message when accessing the streamReader.EndOfStream value.  When this happens, it "restarts" the stream events request "skipping" the already returned events and continues the wait.